\name{RPPairwiseDesign-package}
\alias{RPPairwiseDesign-package}
\alias{RPPairwiseDesign}
\docType{package}
\title{
Resolvable Partially Pairwise Balanced Design and Space-filling Design via Association Scheme.}
\description{
In this package, we apply the (ASC-RPPBD) method on a series of association schemes to construct their associated RPPBD. Moreover, we apply the algorithm (ASC-SF) on the same series of association schemes to obtain their associated space filling design.
Each design is identified by its configutation and its parametrs.
}
\details{
\tabular{ll}{
Package: \tab RPPairwiseDesign\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-12-10\cr
License: \tab GPL-3\cr
}
}
\note{
The Association schemes used in this R-package are :

Rectangular association scheme.

Group divisible association scheme.

Nested group divisible association scheme.

Right angular association scheme

Generalized rectangular right angular association scheme(4)

Generalized rectangular right angular association scheme(5)

Generalized rectangular right angular association scheme(7)
}
\author{
Mohamed Laib, Imane Rezgui and Zebida Gheribi-Aoulmi

Maintainer: Mohamed Laib <laib.med@gmail.com>
}
\references{
Imane Rezgui, Z.Gheribi-Aoulmi and Herve Monod. U-type Designs via New Generalized Partially Balanced Incomplete Block Designs with m = 4, 5 and 7 Associated Classes. Applied Mathematics. to be appear.

Vartak M.N.1955. On an application of Kronecker product of Matrices to Statistical designs. Ann. Math. Stat.,26(420_438).

Imane Rezgui M.LAIB and Z.Gheribi-Aoulmi NEW SERIES OF RESOLVABLE PARTIALLY PAIRWISE BALANCED DESIGNS AND THEIR ASSOCIATED SPACE FILLING DESIGNS
accepted Proceeding on  Afrika Statistika.

Imane Rezgui and Z.Gheribi-Aoulmi New Series of Resolvable Partially Pairwise Balanced and Space Filling Designs via Association Schemes submitted.

S. K. Tharthare, Right angular designs, Ann. Math. Statist. 36 (1963), 1057-1067.

Lokesh Dwivedi. Partially balanced Incomplete block designs. M.Sc. (Agricultural Statistics), Roll No. 4491, I.A.S.R.I., Library Avenue, New Delhi -110 012.
}
